refactor(permission.config): corrected spelling and refactored all files with this import
